services/az-table/requirements.txt
azure-data-tables==12.4.0